# api/swagger.yaml
basePath: /v1
definitions:
...
  domain.Category:
    properties:
      color:
        description: Optional hex color, e.g. "#FF5733"
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      icon:
        description: Optional emoji or icon identifier
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        description: Empty string means root category
        type: string
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.CategoryType'
      updated_at:
        type: string
    type: object
  domain.CategoryType:
    enum:
    - income
    - expense
    - transfer
    type: string
    x-enum-varnames:
    - CategoryTypeIncome
    - CategoryTypeExpense
    - CategoryTypeTransfer
  domain.Role:
    enum:
    - sysadmin
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleSysadmin
    - RoleAdmin
    - RoleMember
  domain.Tenant:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      name:
        type: string
      plan:
        $ref: '#/definitions/domain.TenantPlan'
      updatedAt:
        type: string
    type: object
  domain.TenantPlan:
    enum:
    - free
    - basic
    - premium
    type: string
    x-enum-varnames:
    - TenantPlanFree
    - TenantPlanBasic
    - TenantPlanPremium
...
  domain.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      tenantID:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.CreateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        type: string
      type:
        $ref: '#/definitions/domain.CategoryType'
    required:
    - name
    - type
    type: object
...
  handler.InviteUserRequest:
    properties:
      email:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/domain.Role'
        enum:
        - owner
        - member
    required:
    - email
    - role
    type: object
...
  handler.UpdateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
  handler.UpdateTenantRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: Moolah Household Finance & Investment SaaS API.
  title: Moolah Financial API
  version: "1.0"
paths:
...
  /v1/categories:
    get:
      consumes:
      - application/json
      description: Returns all categories for the current tenant. Can be filtered
        by type or parent ID.
      parameters:
      - description: Category type (income/expense)
        in: query
        name: type
        type: string
      - description: Parent category ULID
        in: query
        name: parent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category for the current tenant.
      parameters:
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Category details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - categories
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a category. This may be blocked if there are transactions
        associated with it.
      parameters:
      - description: Category ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Category deleted
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Returns details of a specific category by ID.
      parameters:
      - description: Category ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get category
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Updates details of an existing category.
      parameters:
      - description: Category ULID
        in: path
        name: id
        required: true
        type: string
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Update fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
  /v1/tenants/me:
    get:
      consumes:
      - application/json
      description: Returns the details of the tenant (household) associated with the
        authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tenant'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tenant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current tenant
      tags:
      - tenants
    patch:
      consumes:
      - application/json
      description: Updates the details of the current tenant (e.g., household name).
      parameters:
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Update fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tenant'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current tenant
      tags:
      - tenants
  /v1/tenants/me/invite:
    post:
      consumes:
      - application/json
      description: Invites a new user (family member) to the current tenant.
      parameters:
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Invite details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.InviteUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (admin only)
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite user
      tags:
      - tenants
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"